Renamed iPAWSDrawable and restructured that a little bit to lay the groundwork for future PAWS extensions.

+++ trunk/src/common/paws/pawslistbox.cpp
@@ -1020,7 +1020,7 @@ void pawsListBox::SortRows()
     {
         rows[i] = sortedRows[i];
         if(sortedRows[i] == selectedrow) //check if the current row is the one which was selected
-            selected = i; //if so update our selected variable
+            selected = (int) i; //if so update our selected variable
     }
 
     delete [] sortedRows;
